Create solicitations Helper
