+++ SIGS/app/controllers/solicitations_controller.rb
@@ -5,6 +5,7 @@
 class SolicitationsController < ApplicationController
   include Schedule
   include PrepareSolicitationsToSave
+  include SolicitationsHelper
   include DateAllocationHelper
   before_action :logged_in?
   before_action :authenticate_not_deg?
@@ -107,33 +108,7 @@ class SolicitationsController < ApplicationController
 
   private
 
-  def validate_status_room(room_solicitation)
-    if !room_solicitation.room_id.nil?
-      1
-    else
-      0
-    end
-  end
-
-  def validade_room_for_approve(room, room_solicitation)
-    if room_solicitation.room_id.nil?
-      room.id
-    else
-      room_solicitation.room_id
-    end
-  end
-
-  def validate_for_save_solicitation(solicitation)
-    solicitation.status = 1
-    return unless solicitation.save
-    flash[:success] = 'Solicitação aprovada com successo'
-    redirect_to solicitations_index_path
-  end
-
-  def pass_to_all_allocation_dates_aux(allocation)
-    pass_to_all_solicitations_helper(allocation)
-  end
-
+  
   def update_room_status(room_solicitation)
     room_solicitation.update(justify: params[:justification],
                              responder_id: current_user,
